feat: use item lang instead of category lang
