+++ package.json
@@ -34,7 +34,7 @@
     "@emotion/server": "11.11.0",
     "@emotion/styled": "11.13.0",
     "@graasp/query-client": "5.0.0",
-    "@graasp/sdk": "4.32.1",
+    "@graasp/sdk": "4.33.0",
     "@graasp/stylis-plugin-rtl": "2.2.0",
     "@graasp/translations": "1.40.0",
     "@graasp/ui": "5.4.0",
