+++ src/components/filters/FilterPopper.tsx
@@ -8,17 +8,11 @@ import {
   Grow,
   Popper,
   Stack,
-  Typography,
   styled,
 } from '@mui/material';
 import { TransitionProps as MUITransitionProps } from '@mui/material/transitions';
 
-import { Category } from '@graasp/sdk';
-
-import {
-  useCategoriesTranslation,
-  useLibraryTranslation,
-} from '../../config/i18n';
+import { useLibraryTranslation } from '../../config/i18n';
 import {
   CLEAR_FILTER_POPPER_BUTTON_ID,
   FILTER_POPPER_ID,
@@ -35,29 +29,28 @@ const StyledPopper = styled(Stack)(() => ({
   boxShadow: '0 2px 15px rgba(0, 0, 0, 0.08)',
 }));
 
-type FilterPopperProps = {
+export type FilterPopperProps = {
   open: boolean;
   anchorEl: HTMLElement | null;
-  options?: Category[];
+  options?: [k: string, v: string][];
   // IDs of selected options.
   selectedOptions: string[];
   onOptionChange: (id: string, newSelected: boolean) => void;
   onClearOptions: () => void;
 };
 
-const FilterPopper = React.forwardRef<HTMLDivElement, FilterPopperProps>(
+export const FilterPopper = React.forwardRef<HTMLDivElement, FilterPopperProps>(
   (
     {
+      options,
       anchorEl,
       onOptionChange,
       open,
-      options,
       selectedOptions,
       onClearOptions,
     },
     ref,
   ) => {
-    const { t: translateCategories } = useCategoriesTranslation();
     const { t } = useLibraryTranslation();
     return (
       <Popper
@@ -73,44 +66,35 @@ const FilterPopper = React.forwardRef<HTMLDivElement, FilterPopperProps>(
           // eslint-disable-next-line react/jsx-props-no-spreading
           <Grow {...TransitionProps}>
             <StyledPopper>
-              {options
-                ?.map((c) => ({ ...c, name: translateCategories(c.name) }))
-                ?.sort(compare)
-                .map((option, idx) => {
-                  const isSelected = selectedOptions.includes(option.id);
-                  return (
-                    <Stack
-                      key={option.id}
-                      id={buildCategoryOptionId(idx)}
-                      direction="row"
-                      alignItems="center"
-                      justifyContent="space-between"
-                      minWidth={200}
-                    >
-                      <FormControl fullWidth>
-                        <FormControlLabel
-                          sx={{
-                            width: '100%',
-                          }}
-                          control={
-                            <Checkbox
-                              checked={isSelected}
-                              onChange={() =>
-                                onOptionChange(option.id, !isSelected)
-                              }
-                            />
-                          }
-                          label={option.name}
-                          labelPlacement="end"
-                        />
-                      </FormControl>
-                    </Stack>
-                  );
-                }) || (
-                <Typography color="#666">
-                  {t(LIBRARY.FILTER_DROPDOWN_NO_CATEGORIES_AVAILABLE)}
-                </Typography>
-              )}
+              {options?.sort(compare).map(([k, v], idx) => {
+                const isSelected = selectedOptions.includes(k);
+                return (
+                  <Stack
+                    key={k}
+                    id={buildCategoryOptionId(idx)}
+                    direction="row"
+                    alignItems="center"
+                    justifyContent="space-between"
+                    minWidth={200}
+                  >
+                    <FormControl fullWidth>
+                      <FormControlLabel
+                        sx={{
+                          width: '100%',
+                        }}
+                        control={
+                          <Checkbox
+                            checked={isSelected}
+                            onChange={() => onOptionChange(k, !isSelected)}
+                          />
+                        }
+                        label={v}
+                        labelPlacement="end"
+                      />
+                    </FormControl>
+                  </Stack>
+                );
+              })}
               <Button
                 id={CLEAR_FILTER_POPPER_BUTTON_ID}
                 variant="outlined"
@@ -127,5 +111,3 @@ const FilterPopper = React.forwardRef<HTMLDivElement, FilterPopperProps>(
     );
   },
 );
-
-export default FilterPopper;
