+++ src/langs/constants.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export const LIBRARY = {
   GRAASP_LIBRARY: 'GRAASP_LIBRARY',
   GRAASP_LIBRARY_DESCRIPTION: 'GRAASP_LIBRARY_DESCRIPTION',
@@ -56,7 +55,7 @@ export const LIBRARY = {
   COLLECTION_CATEGORIES_TITLE: 'COLLECTION_CATEGORIES_TITLE',
   CREATE_BUTTON_TOOLTIP: 'CREATE_BUTTON_TOOLTIP',
   COLLECTION_TAGS_TITLE: 'COLLECTION_TAGS_TITLE',
-  COLLECTION_LANGUAGES_TITLE: 'COLLECTION_LANGUAGES_TITLE',
+  COLLECTION_LANGUAGE_TITLE: 'COLLECTION_LANGUAGE_TITLE',
   PROFILE_BUTTON: 'PROFILE_BUTTON',
   COLLECTIONS_COUNT_MESSAGE: 'COLLECTIONS_COUNT_MESSAGE',
   HOME_TITLE: 'HOME_TITLE',
