+++ src/langs/en.json
@@ -54,7 +54,7 @@
   "COLLECTION_CATEGORIES_TITLE": "Categories",
   "CREATE_BUTTON_TOOLTIP": "Create your own collection",
   "COLLECTION_TAGS_TITLE": "Tags",
-  "COLLECTION_LANGUAGES_TITLE": "Languages",
+  "COLLECTION_LANGUAGE_TITLE": "Language",
   "PROFILE_BUTTON": "Account",
   "COLLECTIONS_COUNT_MESSAGE_one": "{{count}} collection available",
   "COLLECTIONS_COUNT_MESSAGE_other": "{{count}} collections available",
