+++ src/langs/es.json
@@ -54,7 +54,6 @@
   "COLLECTION_CATEGORIES_TITLE": "Categorías",
   "CREATE_BUTTON_TOOLTIP": "Crea tu propia colección",
   "COLLECTION_TAGS_TITLE": "Etiquetas",
-  "COLLECTION_LANGUAGES_TITLE": "Idiomas",
   "PROFILE_BUTTON": "Cuenta",
   "COLLECTIONS_COUNT_MESSAGE_one": "{{count}} colección disponible",
   "COLLECTIONS_COUNT_MESSAGE_other": "{{count}} colecciones disponibles",
