+++ src/langs/it.json
@@ -54,7 +54,6 @@
   "COLLECTION_CATEGORIES_TITLE": "Categorie",
   "CREATE_BUTTON_TOOLTIP": "Crea la tua collezione",
   "COLLECTION_TAGS_TITLE": "Tag",
-  "COLLECTION_LANGUAGES_TITLE": "Le lingue",
   "PROFILE_BUTTON": "Account",
   "COLLECTIONS_COUNT_MESSAGE_one": "{{count}} collezione disponibile",
   "COLLECTIONS_COUNT_MESSAGE_other": "{{count}} collezioni disponibili",
